+++ docs/src/release_notes.rst
@@ -8,6 +8,7 @@ Unreleased
 Added
 *****
 - :ref:`conversions` - Ignore GET parameters in :func:`from_url` (:issue:`226`)
+
 Fixed
 *****
 - :ref:`client` - document new behavior of track markets (:issue:`228`)